# Conflicts:
#	src/Connector.DataLake.Common/Connector/DataLakeStorageClient.cs

+++ src/Connector.DataLake.Common/Connector/DataLakeStorageClient.cs
@@ -41,6 +41,11 @@ internal class DataLakeStorageClient : IStorageClient
     {
         var fileClient = await GetFileClientAsync(filePath, createDirectoryIfNotExists: false);
 
+        if (fileClient == null)
+        {
+            return;
+        }
+
         await fileClient.DeleteIfExistsAsync();
     }
 
@@ -80,6 +85,11 @@ internal class DataLakeStorageClient : IStorageClient
     public async Task<FileMetadata> GetFileMetadataAsync(FilePath filePath)
     {
         var fileClient = await GetFileClientAsync(filePath, createDirectoryIfNotExists: false);
+
+        if (fileClient == null)
+        {
+            return null;
+        }
 
         if (!await fileClient.ExistsAsync())
         {
